+++ tests/GridView/Column/RadioColumnTest.php
@@ -130,6 +130,25 @@ final class RadioColumnTest extends TestCase
         );
     }
 
+    public function testHeaderAttributesWithoutHeader(): void
+    {
+        $html = $this->createGridView([['id' => 1]])
+            ->headerCellAttributes(['class' => 'header-cell-class'])
+            ->columns(
+                new RadioColumn(
+                    headerAttributes: ['style' => 'width: 1%'],
+                ),
+            )
+            ->render();
+
+        $this->assertStringContainsString(
+            <<<HTML
+            <th class="header-cell-class" style="width: 1%">&nbsp;</th>
+            HTML,
+            $html,
+        );
+    }
+
     public function testBodyAttributes(): void
     {
         $html = $this->createGridView([['id' => 1]])